pyproject.toml
[tool.poetry]
name = "kafka-test-task"
version = "0.1.0"
description = ""
authors = ["Nick Smirnov <na.smirnov@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
